% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChargaffTable.R
\docType{data}
\name{ChargaffTable}
\alias{ChargaffTable}
\title{Nucleotide frequencies discovered by Chargaff}
\format{
matrix that contains the mass frequency percentages for each nucleotide
\describe{
\item{\code{1:4}}{chr nucleotides A,T,C,G}
\item{\code{1:8}}{chr organism names}
\item{\code{1:8},\code{1:4}}{num the percentage pertaining to the nucleotide and organism}
}
}
\source{
\url{https://web.stanford.edu/class/bios221/book/02-chap.html}
}
\usage{
ChargaffTable
}
\description{
percentages of the mass present in different organisms for each of the nucleotides, not the measurements themselves.
}
\keyword{datasets}
