% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deuteranopia.R
\docType{data}
\name{Deuteranopia}
\alias{Deuteranopia}
\title{Red-green color blindness data}
\format{
matrix that contains a two-way table
\describe{
\item{\code{1:4}}{chr Deute and NonDeute}
\item{\code{1:2}}{chr Men and Women}
\item{\code{1:2},\code{1:2}}{num results of the two-way table}
}
}
\source{
\url{https://web.stanford.edu/class/bios221/book/02-chap.html}
}
\usage{
Deuteranopia
}
\description{
Deuteranopia is a form of red-green color blindness due to the fact that medium wavelength sensitive cones (green) are missing. A deuteranope can only distinguish 2 to 3 different hues, whereas somebody with normal vision sees 7 different hues. A survey for this type of color blindness in human subjects produced a two-way table crossing color blindness and sex.
}
\keyword{datasets}
