% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e100.R
\docType{data}
\name{e100}
\alias{e100}
\title{ELISA assay data from Modern Statistics for Modern Biology}
\format{
A numeric vector containing ELISA assay data for 50 patients tallied across 100 positions:
\describe{
\item{\code{1-100}}{dbl assay data for 50 patients tallied at each of the 100 positions}
}
}
\source{
\url{https://web.stanford.edu/class/bios221/book/01-chap.html}
}
\usage{
e100
}
\description{
ELISA assay data from Modern Statistics for Modern Biology
}
\keyword{datasets}
