% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{function that runs the install.packages function for required libraries}
\usage{
install(install_all = TRUE)
}
\arguments{
\item{install_all}{defaults to TRUE to allow the packages to be installed.}
}
\value{
nothing
}
\description{
function that runs the install.packages function for required libraries
}
\examples{
# install(install_all = TRUE)
}
